ticket_generator/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
kafka-python==2.2.7
    # via 09-spark-redpanda (pyproject.toml)
python-dotenv==1.1.0
    # via 09-spark-redpanda (pyproject.toml)
